+++ Thunderstore/Package Files/CHANGELOG.md
@@ -6,13 +6,13 @@ Item highlights being survivor colored now use a better method
 Changed item highlight texture to a cleaner one
 
 Tweaked the TAB inventory menu item displays
-- Item icons are more spaced out (they're super cramped in vanilla at high item counts)
+- Item icons are more spaced out (there's practically no space between icons in vanilla at high item counts)
 - Item icons display should take up more height in a player's inventories menu section
 
 Fixed an extra background detail for the item icon in the inspect not being removed
 - Only noticed it recently as the color difference for this detail and the rest of the inspect panel is *very* slight
 
-Centered player names on the inventory menu
+(Hopefully) Centered player names on the inventory menu
 
 Made HP bar text smoother